Move standard deck functionality into deck class
