+++ lib/deck.rb
@@ -1,8 +1,24 @@
 class Deck
   attr_reader :cards
 
-  def initialize(cards)
+  def initialize(cards = [])
     @cards = cards
+    @suit_array = [:club, :diamond, :heart, :spade]
+    @rank_values = {
+      2 => '2',
+      3 => '3',
+      4 => '4',
+      5 => '5',
+      6 => '6',
+      7 => '7',
+      8 => '8',
+      9 => '9',
+      10 => '10',
+      11 => 'Jack',
+      12 => 'Queen',
+      13 => 'King',
+      14 => 'Ace'
+    }
   end
 
   def rank_of_card_at(index)
@@ -25,6 +41,19 @@ class Deck
 
   def add_card(new_card)
     cards << new_card
+  end
+
+  def populate_standard_deck
+    @suit_array.each do |suit|
+      @rank_values.each do |rank, value|
+        new_card = Card.new(suit, value, rank)
+        add_card(new_card)
+      end
+    end
+  end
+
+  def shuffle_deck
+    cards.shuffle
   end
 
 end
